+++ src/logger.ts
@@ -1,5 +1,5 @@
 import { window } from "vscode"
 
-const logger = window.createOutputChannel("Compiler Messages", "stan");
+const logger = window.createOutputChannel("Stan Extension");
 
 export default logger;